added ConnectionPool, added UserDao, method getAllUsers in it and tests for them

+++ src/main/java/models/implementations/Order.java
@@ -5,11 +5,11 @@ import models.IOrder;
 
 public class Order implements IOrder {
     private OrderStatus status;
-    private int price;
+    private double price;
     private final User user;
     private final String car;
 
-    public Order(int price, String car, User user, OrderStatus status) {
+    public Order(double price, String car, User user, OrderStatus status) {
         this.status = status;
         this.price = price;
         this.user = user;
